chg: Renamed data_pipeline more specifically to training_pipeline
